bbdev: extend range of allocation functions
Realigning the argument to unsigned int to
align with number support by underlying
rte_mempool_get_bulk function.

Acked-by: Hemant Agrawal <hemant.agrawal@nxp.com>

+++ lib/bbdev/rte_bbdev_op.h
@@ -981,7 +981,7 @@ rte_bbdev_op_pool_create(const char *name, enum rte_bbdev_op_type type,
  */
 static inline int
 rte_bbdev_enc_op_alloc_bulk(struct rte_mempool *mempool,
-		struct rte_bbdev_enc_op **ops, uint16_t num_ops)
+		struct rte_bbdev_enc_op **ops, unsigned int num_ops)
 {
 	struct rte_bbdev_op_pool_private *priv;
 
@@ -1012,7 +1012,7 @@ rte_bbdev_enc_op_alloc_bulk(struct rte_mempool *mempool,
  */
 static inline int
 rte_bbdev_dec_op_alloc_bulk(struct rte_mempool *mempool,
-		struct rte_bbdev_dec_op **ops, uint16_t num_ops)
+		struct rte_bbdev_dec_op **ops, unsigned int num_ops)
 {
 	struct rte_bbdev_op_pool_private *priv;
 
@@ -1044,7 +1044,7 @@ rte_bbdev_dec_op_alloc_bulk(struct rte_mempool *mempool,
 __rte_experimental
 static inline int
 rte_bbdev_fft_op_alloc_bulk(struct rte_mempool *mempool,
-		struct rte_bbdev_fft_op **ops, uint16_t num_ops)
+		struct rte_bbdev_fft_op **ops, unsigned int num_ops)
 {
 	struct rte_bbdev_op_pool_private *priv;
